# .safira/tmp/786adc15-ec20-4f92-8a86-cebdf8e4c6bb/src/gen/java/br/com/vertigo/employee/api/HerosApi.java
package br.com.vertigo.employee.api;


import javax.validation.constraints.*;
import javax.ws.rs.*;
import javax.ws.rs.core.Response;

@Path("/heros")
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaJAXRSSpecServerCodegen")
public interface HerosApi {

  @GET
  @Produces({"application/json"})
  Response herosGet();

  @GET
  @Path("/{id}")
  @Produces({"application/json"})
  Response herosIdGet(@PathParam("id") String id);
}