# .safira/tmp/b9e25743-4b30-4933-8ece-281ef6fd1ab8/src/gen/java/br/com/vertigo/employee/api/HerosApi.java
package br.com.vertigo.employee.api;


import javax.validation.constraints.*;
import javax.ws.rs.*;
import javax.ws.rs.core.Response;

@Path("/heros")
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaJAXRSSpecServerCodegen")
public interface HerosApi {

  @GET
  @Produces({"application/json"})
  Response herosGet();

  @GET
  @Path("/{id}")
  @Produces({"application/json"})
  Response herosIdGet(@PathParam("id") Integer id);
}